Authenticate with User ID

+++ lsapp/app/Http/Controllers/ListingController.php
@@ -51,6 +51,8 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
+        $formFields['user_id'] = auth()->id();
+
         Listing::create($formFields);
 
         return redirect('/')->with('message', 'Listing created successfully');
@@ -64,6 +66,13 @@ class ListingController extends Controller
 
     // Update listing data
     public function update(Request $request, Listing $listing) {
+
+        // Make sure logged in user is owner
+
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized action');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required'],
@@ -88,8 +97,20 @@ class ListingController extends Controller
     // Delete Listing
 
     public function destroy(Listing $listing) {
+
+        // Make sure logged in user is owner
+
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized action');
+        }
+ 
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted Successfully');
+    }
+
+    // Manage Listings
+    public function manage() {
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 
 }
